main/requirements.txt
click
Flask
grpcio
itsdangerous
Jinja2
MarkupSafe
six
Werkzeug
google
protobuf
